backend changes v1

+++ backend/controllers/group-chat/group-chat.js
@@ -1,4 +1,6 @@
 const GroupChatModal = require('../../models/chat/group-chat-modal/group-chat-model')
+const chatHistory = require('../../models/chat/chat-history/chat-history-modal')
+
 
 const createGroupChat = async (req, res) => {
     const { group_name, group_description, group_admin, participant, admin_id } = req.body
@@ -11,9 +13,7 @@ const createGroupChat = async (req, res) => {
         })
         if (makeChatGroup) {
             const findCreatedGroup = await GroupChatModal.findById(makeChatGroup._id)
-            const filterParticipants = findCreatedGroup.participant
-                .filter(part => part._id !== admin_id && part._id !== null)
-                .map(part => part._id);
+            const filterParticipants = findCreatedGroup.participant.map(part => part._id);
 
             return res.status(201).send({ message: "Chat Group Created", participant: filterParticipants, groupData: findCreatedGroup })
         } else {
@@ -100,10 +100,25 @@ const deleteGroupChat = async (req, res) => {
         return res.status(400).send({ message: 'id not found' })
     }
     try {
-        const groups = await GroupChatModal.findByIdAndRemove(_id)
-        if (groups) return res.status(200).send({ message: 'Group Deleted' })
-        else return res.status(404).send({ message: 'Group not found' })
-
+        const findGroup = await GroupChatModal.findById(_id)
+        if (findGroup) {
+            const participant = findGroup.participant.map(part => part._id);
+            if (participant.length > 0) {
+                for (let i = 0; i <= participant.length - 1; i++) {
+                    const findChatHistory = await chatHistory.findOne({ userId: participant[i] })
+                    if (findChatHistory) {
+                        const filterH = findChatHistory.chated_persons.filter(cp => cp.user_id !== findGroup._id)
+                        findChatHistory.chated_persons = filterH
+                        await findChatHistory.save()
+                    }
+                }
+                const groups = await GroupChatModal.findByIdAndRemove(_id)
+                if (groups) return res.status(200).send({ message: 'Group Deleted' })
+            }
+        }
+        else {
+            return res.status(404).send({ message: 'Group not found' })
+        }
     } catch (error) {
         return res.status(500).send({ message: "Internal Server Error" })
     }
